requirements.txt
# Core framework
fastapi==0.135.2
starlette==1.0.0
uvicorn>=0.29.0

# Async / HTTP
anyio==4.13.0
httpcore==1.0.9
httpx==0.28.1
h11==0.16.0
sniffio==1.3.1

# Pydantic / settings
pydantic[email]==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
annotated-types==0.7.0
annotated-doc==0.0.4
typing-inspection==0.4.2
typing_extensions==4.15.0

# Database
SQLAlchemy==2.0.48

# Auth — NEW
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0
python-multipart>=0.0.9   # required for OAuth2 form login
email-validator>=2.1.0    # required for Pydantic EmailStr

# OpenAI
openai==2.30.0
jiter==0.13.0

# Misc
certifi==2026.2.25
distro==1.9.0
idna==3.11
python-dotenv==1.2.2
tqdm==4.67.3